[RTF-1414] Workspace Sreate Strategy V1
