--- source/Relativity.Testing.Framework.Api/Strategies/ResourcePools/IResourcePoolGetEligibleStrategy.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using Relativity.Testing.Framework.Models;
-
-namespace Relativity.Testing.Framework.Api.Strategies
-{
-	internal interface IResourcePoolGetEligibleStrategy
-	{
-		List<NamedArtifact> GetEligibleResourcePools();
-	}
-}
